Update template to use the latest version of fabric loom (#1)
* wipe

yes, we doing this oh boyo

* add license and readme

* introduce gitignore, recreate template + tasks

need to run some quick tests

* fix update task, bump to v2 fabric api

* reformat

once and done

* import build workflow from fabric

small changes to use temurin instead, and to prevent daemon creation by gradle (we dont need it on a workflow).

* add fix for spaces in props file

task can support if there are spaces or no spaces between the equal sign when assigning the prop.

* switch up hardcoded version to the sourceCompat, and add notice

we shouldn't change the java version in most cases. It will be many many versions imo before java 20 will be used for MC, so until then its best that unsuspecting people do not mess with it.